requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
schedule==1.2.0
python-multipart==0.0.6

